Some small style fixes and adjustments
