+++ lib/puppet/type/aem_content.rb
